+++ Soko2019/src/assignment4/Timeable.java
@@ -2,6 +2,6 @@ package assignment4;
 
 public interface Timeable {
 	
-	public void setTimer(long timer);
+	public void setTimer(int timer);
 	
 }
